requirements.txt
# Dependencies required for the project

google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
openai
python-dotenv
google-auth
pickle-mixin
